[~] Rename kollabStorage to kstorage
