+++ apps/kstorage/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class KollabstorageConfig(AppConfig):
-    name = 'apps.kollabStorage'
+    name = 'apps.kstorage'
