+++ apps/kstorage/views.py
